+++ src/RestApi/Device/PeerRecovery/DevicePeerRecovery.ts
@@ -3,6 +3,7 @@ import { IDeviceActionLog } from '../Telemetry/IDeviceTelemetry';
 
 export interface IDevicePeerRecovery extends IDeviceActionLog {
 	enabled: boolean;
+	autoEnableTimeoutMs?: number;
 }
 
 export default class DevicePeerRecovery implements IDevicePeerRecovery {
@@ -12,6 +13,7 @@ export default class DevicePeerRecovery implements IDevicePeerRecovery {
 	public readonly succeededAt: IDevicePeerRecovery['succeededAt'];
 	public readonly failedAt: IDevicePeerRecovery['failedAt'];
 	public readonly enabled: IDevicePeerRecovery['enabled'];
+	public readonly autoEnableTimeoutMs: IDevicePeerRecovery['autoEnableTimeoutMs'];
 
 	constructor(data: IDevicePeerRecovery) {
 		fillDataToEntity(this, data);